refactor: put trainer and export code with their owners
trainer_run.py becomes train_cfg_io.py: it reads the training run's config
and artifacts, and the old name read as "run the trainer". write_lr_history
moves into the magic trainer that calls it, ending magic's import of
approx_unrolling; ensure_exported and EXPORT_DIRNAME move next to
export_checkpoints in utils.trainer_export.
